+++ app/client/src/pages/Editor/gitSync/Tabs/GitConnection.tsx
@@ -12,8 +12,6 @@ import {
   REMOTE_URL,
   REMOTE_URL_INFO,
   createMessage,
-  DEPLOY_KEY_USAGE_GUIDE_MESSAGE,
-  DEPLOY_KEY_TITLE,
   REMOTE_URL_INPUT_PLACEHOLDER,
   CONNECTING_REPO,
 } from "constants/messages";
@@ -23,24 +21,19 @@ import UserGitProfileSettings from "../components/UserGitProfileSettings";
 import { AUTH_TYPE_OPTIONS } from "../constants";
 import { Colors } from "constants/Colors";
 import Button, { Category, Size } from "components/ads/Button";
-import { useGitConnect, useSSHKeyPair } from "../hooks";
-import { ReactComponent as KeySvg } from "assets/icons/ads/key-2-line.svg";
-import { ReactComponent as CopySvg } from "assets/icons/ads/file-copy-line.svg";
-import { ReactComponent as TickSvg } from "assets/images/tick.svg";
+import { useGitConnect, useSSHKeyPair, useUserGitConfig } from "../hooks";
 import { Toaster } from "components/ads/Toast";
 import { Variant } from "components/ads/common";
 import { useDispatch, useSelector } from "react-redux";
 import copy from "copy-to-clipboard";
 import { getCurrentAppGitMetaData } from "selectors/applicationSelectors";
 import Text, { TextType } from "components/ads/Text";
-import { getGlobalGitConfig } from "selectors/gitSyncSelectors";
+
 import {
   fetchGlobalGitConfigInit,
   fetchLocalGitConfigInit,
   updateLocalGitConfigInit,
 } from "actions/gitSyncActions";
-import TooltipComponent from "components/ads/Tooltip";
-import { getLocalGitConfig } from "selectors/gitSyncSelectors";
 import { emailValidator } from "components/ads/TextInput";
 import { isEqual } from "lodash";
 import {
@@ -59,6 +52,7 @@ import Statusbar from "pages/Editor/gitSync/components/Statusbar";
 import ScrollIndicator from "components/ads/ScrollIndicator";
 import GitSyncError from "../components/GitError";
 import { ReduxActionErrorTypes } from "constants/ReduxActionConstants";
+import DeployedKeyUI from "../components/DeployedKeyUI";
 
 export const UrlOptionContainer = styled.div`
   display: flex;
@@ -84,85 +78,6 @@ const ButtonContainer = styled.div<{ topMargin: number }>`
   margin-top: ${(props) => `${props.theme.spaces[props.topMargin]}px`};
 `;
 
-const Icon = styled.span<{
-  size: string;
-  color: string;
-  marginOffset?: number;
-  hoverColor: string;
-}>`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin: ${(props) => `${props.theme.spaces[props.marginOffset || 0]}px`};
-  cursor: pointer;
-  svg {
-    width: ${(props) => props.size};
-    height: ${(props) => props.size};
-    path {
-      fill: ${(props) => props.color};
-    }
-  }
-  &:hover {
-    svg {
-      path {
-        fill: ${(props) => props.hoverColor};
-      }
-    }
-  }
-`;
-
-const DeployedKeyContainer = styled.div<{ $marginTop: number }>`
-  margin-top: ${(props) => `${props.theme.spaces[props.$marginTop]}px`};
-  margin-bottom: 8px;
-  height: 50px;
-  width: calc(100% - 30px);
-  background-color: ${Colors.Gallery};
-  padding: ${(props) =>
-    `${props.theme.spaces[3]}px ${props.theme.spaces[4]}px`};
-`;
-
-const FlexColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 90%;
-  margin-left: ${(props) => `${props.theme.spaces[4]}px`};
-`;
-
-const FlexRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-`;
-
-const Flex = styled.div`
-  display: flex;
-`;
-
-const LabelText = styled.span`
-  font-size: 14px;
-  color: ${Colors.CODE_GRAY};
-`;
-
-const KeyText = styled.span`
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  width: 100%;
-  font-size: 10px;
-  font-weight: 400;
-  text-transform: uppercase;
-  color: ${Colors.CODE_GRAY};
-`;
-
-const LintText = styled.a`
-  :hover {
-    text-decoration: none;
-    color: ${Colors.CRUSTA};
-  }
-  color: ${Colors.CRUSTA};
-  cursor: pointer;
-`;
-
 const Container = styled.div`
   display: flex;
   flex-direction: column;
@@ -175,12 +90,6 @@ const Container = styled.div`
   &::-webkit-scrollbar {
     width: 0px;
   }
-`;
-
-const TooltipWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
 `;
 
 const RemoteUrlInfoWrapper = styled.div`
@@ -205,6 +114,11 @@ const StickyMenuWrapper = styled.div`
 const selectedAuthType = AUTH_TYPE_OPTIONS[0];
 const HTTP_LITERAL = "https";
 
+const SSH_INIT_FORMAT_REGEX = new RegExp(/^(ssh|.+@).*/);
+
+const remoteUrlIsInvalid = (value: string) =>
+  value.startsWith(HTTP_LITERAL) || !SSH_INIT_FORMAT_REGEX.test(value);
+
 type Props = {
   isImport?: boolean;
 };
@@ -219,40 +133,15 @@ function GitConnection({ isImport }: Props) {
   const isFetchingLocalGitConfig = useSelector(getIsFetchingLocalGitConfig);
   const [triedSubmit, setTriedSubmit] = useState(false);
 
-  const globalGitConfig = useSelector(getGlobalGitConfig);
-  const localGitConfig = useSelector(getLocalGitConfig);
-  const isLocalConfigDefined = !!(
-    localGitConfig.authorEmail || localGitConfig.authorName
-  );
-
-  const isGlobalConfigDefined = !!(
-    globalGitConfig.authorEmail || globalGitConfig.authorName
-  );
-
   const dispatch = useDispatch();
 
-  const getInitGitConfig = useCallback(() => {
-    let initialAuthInfo = {
-      authorName: "",
-      authorEmail: "",
-    };
-
-    if (isGlobalConfigDefined) {
-      initialAuthInfo = {
-        authorName: globalGitConfig.authorName || "",
-        authorEmail: globalGitConfig.authorEmail || "",
-      };
-    }
-    // when local config is defined we will only show local config
-    if (isLocalConfigDefined) {
-      initialAuthInfo = {
-        authorName: localGitConfig.authorName || "",
-        authorEmail: localGitConfig.authorEmail || "",
-      };
-    }
-
-    return initialAuthInfo;
-  }, [globalGitConfig, localGitConfig]);
+  const {
+    getInitGitConfig,
+    globalGitConfig,
+    isGlobalConfigDefined,
+    isLocalConfigDefined,
+    localGitConfig,
+  } = useUserGitConfig();
 
   const initialAuthorInfoRef = useRef(getInitGitConfig());
 
@@ -274,7 +163,6 @@ function GitConnection({ isImport }: Props) {
 
   const {
     deployKeyDocUrl,
-    // failedGeneratingSSHKey,
     fetchingSSHKeyPair,
     fetchSSHKeyPair,
     generateSSHKey,
@@ -355,12 +243,14 @@ function GitConnection({ isImport }: Props) {
           dispatch(updateLocalGitConfigInit(authorInfo));
         }
       } else {
-        connectToGit({
-          remoteUrl,
-          gitProfile: authorInfo,
-          isImport,
-          isDefaultProfile: useGlobalConfig,
-        });
+        if (!isInvalidRemoteUrl) {
+          connectToGit({
+            remoteUrl,
+            gitProfile: authorInfo,
+            isImport,
+            isDefaultProfile: useGlobalConfig,
+          });
+        }
       }
     } else {
       Toaster.show({
@@ -388,16 +278,7 @@ function GitConnection({ isImport }: Props) {
     setRemoteUrl(value);
   };
 
-  const remoteUrlIsInvalid = useCallback(
-    (value: string) => value.startsWith(HTTP_LITERAL),
-    [],
-  );
-
   const submitButtonDisabled = useMemo(() => {
-    // const isAuthorInfoEmpty = !authorInfo.authorEmail || !authorInfo.authorName;
-    // const isAuthorEmailInvalid = !emailValidator(authorInfo.authorEmail)
-    //   .isValid;
-
     const isAuthInfoUpdated = isAuthorInfoUpdated();
     let buttonDisabled = false;
     if (isGitConnected) {
@@ -441,6 +322,7 @@ function GitConnection({ isImport }: Props) {
   }, [setUseGlobalConfig, useGlobalConfig]);
 
   const scrollWrapperRef = React.createRef<HTMLDivElement>();
+
   useEffect(() => {
     if (gitError.message && scrollWrapperRef.current) {
       const top = scrollWrapperRef.current.scrollHeight;
@@ -500,52 +382,12 @@ function GitConnection({ isImport }: Props) {
             </ButtonContainer>
           )
         ) : (
-          <>
-            <FlexRow>
-              <DeployedKeyContainer $marginTop={8}>
-                <FlexRow>
-                  <Flex>
-                    <KeySvg />
-                  </Flex>
-
-                  <FlexColumn>
-                    <LabelText>{createMessage(DEPLOY_KEY_TITLE)}</LabelText>
-                    <KeyText>{SSHKeyPair}</KeyText>
-                  </FlexColumn>
-                </FlexRow>
-              </DeployedKeyContainer>
-              {showCopied ? (
-                <Icon
-                  color={Colors.GREEN}
-                  hoverColor={Colors.GREEN}
-                  marginOffset={4}
-                  size="16px"
-                >
-                  <TickSvg />
-                </Icon>
-              ) : (
-                <TooltipWrapper>
-                  <TooltipComponent content="Copy Key">
-                    <Icon
-                      color={Colors.DARK_GRAY}
-                      hoverColor={Colors.GRAY2}
-                      marginOffset={3}
-                      onClick={copyToClipboard}
-                      size="22px"
-                    >
-                      <CopySvg />
-                    </Icon>
-                  </TooltipComponent>
-                </TooltipWrapper>
-              )}
-            </FlexRow>
-            <Text color={Colors.GREY_9} type={TextType.P3}>
-              {createMessage(DEPLOY_KEY_USAGE_GUIDE_MESSAGE)}
-              <LintText href={deployKeyDocUrl} target="_blank">
-                &nbsp;LEARN MORE
-              </LintText>
-            </Text>
-          </>
+          <DeployedKeyUI
+            SSHKeyPair={SSHKeyPair}
+            copyToClipboard={copyToClipboard}
+            deployKeyDocUrl={deployKeyDocUrl}
+            showCopied={showCopied}
+          />
         )}
       </Section>
