+++ app/client/src/pages/Editor/gitSync/hooks.ts
@@ -5,6 +5,12 @@ import { connectToGitInit } from "actions/gitSyncActions";
 import { ConnectToGitPayload } from "api/GitSyncAPI";
 import { getCurrentApplication } from "selectors/applicationSelectors";
 import { DOCS_BASE_URL } from "constants/ThirdPartyConstants";
+import {
+  getGlobalGitConfig,
+  getLocalGitConfig,
+  getIsGlobalConfigDefined,
+  getIsLocalConfigDefined,
+} from "selectors/gitSyncSelectors";
 
 export const useSSHKeyPair = () => {
   // As SSHKeyPair fetching and generation is only done only for GitConnection part,
@@ -117,5 +123,43 @@ export const useGitConnect = () => {
     isConnectingToGit,
     gitError,
     connectToGit,
+  };
+};
+
+export const useUserGitConfig = () => {
+  const globalGitConfig = useSelector(getGlobalGitConfig);
+  const localGitConfig = useSelector(getLocalGitConfig);
+  const isLocalConfigDefined = useSelector(getIsLocalConfigDefined);
+  const isGlobalConfigDefined = useSelector(getIsGlobalConfigDefined);
+
+  const getInitGitConfig = useCallback(() => {
+    let initialAuthInfo = {
+      authorName: "",
+      authorEmail: "",
+    };
+
+    if (isGlobalConfigDefined) {
+      initialAuthInfo = {
+        authorName: globalGitConfig.authorName || "",
+        authorEmail: globalGitConfig.authorEmail || "",
+      };
+    }
+    // when local config is defined we will only show local config
+    if (isLocalConfigDefined) {
+      initialAuthInfo = {
+        authorName: localGitConfig.authorName || "",
+        authorEmail: localGitConfig.authorEmail || "",
+      };
+    }
+
+    return initialAuthInfo;
+  }, [globalGitConfig, localGitConfig]);
+
+  return {
+    getInitGitConfig,
+    globalGitConfig,
+    isGlobalConfigDefined,
+    isLocalConfigDefined,
+    localGitConfig,
   };
 };
