+++ app/client/src/selectors/gitSyncSelectors.tsx
@@ -51,6 +51,18 @@ export const getGlobalGitConfig = (state: AppState) =>
 export const getLocalGitConfig = (state: AppState) =>
   state.ui.gitSync.localGitConfig;
 
+export const getIsLocalConfigDefined = createSelector(
+  getLocalGitConfig,
+  (localGitConfig) =>
+    !!(localGitConfig.authorEmail || localGitConfig.authorName),
+);
+
+export const getIsGlobalConfigDefined = createSelector(
+  getGlobalGitConfig,
+  (globalGitConfig) =>
+    !!(globalGitConfig.authorEmail || globalGitConfig.authorName),
+);
+
 export const getIsFetchingGlobalGitConfig = (state: AppState) =>
   state.ui.gitSync.isFetchingGitConfig;
